Rearrange decorators
