+++ types.go
@@ -1,13 +1,86 @@
 package lua
 
 import (
-	"github.com/krakendio/binder"
+	"errors"
+	"sort"
+	"strconv"
+
+	glua "github.com/yuin/gopher-lua"
 )
 
-func RegisterNil(b *binder.Binder) {
-	tab := b.Table("luaNil")
-	tab.Static("new", func(c *binder.Context) error {
-		c.Push().Data(nil, "luaNil")
-		return nil
-	})
+type Table struct {
+	Data map[string]interface{}
+}
+
+type List struct {
+	Data []interface{}
+}
+
+type NativeValue = glua.LValue
+type NativeNumber = glua.LNumber
+type NativeString = glua.LString
+type NativeBool = glua.LBool
+type NativeUserData = glua.LUserData
+type NativeTable = glua.LTable
+
+func ParseToTable(k, v NativeValue, acc map[string]interface{}) {
+	switch v.Type() {
+	case glua.LTString:
+		acc[k.String()] = v.String()
+	case glua.LTBool:
+		acc[k.String()] = glua.LVAsBool(v)
+	case glua.LTNumber:
+		f := float64(v.(NativeNumber))
+		if f == float64(int64(v.(NativeNumber))) {
+			acc[k.String()] = int(v.(NativeNumber))
+		} else {
+			acc[k.String()] = f
+		}
+	case glua.LTUserData:
+		userV := v.(*NativeUserData)
+		if userV.Value == nil {
+			acc[k.String()] = nil
+		} else {
+			switch v := userV.Value.(type) {
+			case *Table:
+				acc[k.String()] = v.Data
+			case *List:
+				acc[k.String()] = v.Data
+			}
+		}
+	case glua.LTTable:
+		res := map[string]interface{}{}
+		v.(*NativeTable).ForEach(func(k, v NativeValue) {
+			ParseToTable(k, v, res)
+		})
+		// Check if all the keys are integers and convert to array
+		acc[k.String()] = res
+		t, err := tryConvertToArray(res)
+		if err == nil {
+			acc[k.String()] = t
+		}
+	}
+}
+
+func tryConvertToArray(input map[string]interface{}) ([]interface{}, error) {
+	keys := make([]int, 0, len(input))
+	values := map[int]interface{}{}
+
+	for k, v := range input {
+		ik, err := strconv.Atoi(k)
+		if err != nil {
+			return nil, errors.New("non-integer key, cannot convert")
+		}
+		keys = append(keys, ik)
+		values[ik] = v
+	}
+
+	sort.Ints(keys)
+
+	result := []interface{}{}
+	for _, k := range keys {
+		result = append(result, values[k])
+	}
+
+	return result, nil
 }
